Move tables test files to tables folder
