+++ Database/lib/tables/attribute.js
