+++ Database/lib/tables/course.js
